Added new and updated functions, added folders
Some improvements were made to the user_ names and levels function, including the implementation of AR.
Added the generate_list function to make a random list for each player.
Added a level_report function, it is unfinished.
Added a main_menu function so the player can do different actions within the game.
Added a game function where every other function needed for the game will be called.
Added the markers and images folders that will be used for the AR feature.

Co-Authored-By: Mariangel2828 <162812751+Mariangel2828@users.noreply.github.com>